requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
pinecone-client>=3.0.0
PyPDF2>=3.0.0
python-dotenv>=1.0.0
requests>=2.31.0
tqdm>=4.66.0
torch>=2.0.0
